+++ murrelet_draw/src/sequencers.rs
@@ -1,7 +1,9 @@
 use glam::*;
 use itertools::Itertools;
 use murrelet_common::*;
-use murrelet_livecode::unitcells::{UnitCellContext, UnitCellCreator, UnitCellExprWorldContext};
+use murrelet_livecode::unitcells::{
+    UnitCellContext, UnitCellCreator, UnitCellEvalContext, UnitCellExprWorldContext,
+};
 use murrelet_livecode_derive::*;
 
 const REFERENCE_SIZE: f32 = 100.0;
@@ -13,11 +15,11 @@ pub enum Sequencer {
     Hex(SimpleHexSequence),
 }
 impl UnitCellCreator for Sequencer {
-    fn to_unit_cell_ctxs(&self) -> Vec<UnitCellContext> {
+    fn to_unit_cell_ctxs(&self, ctx: &UnitCellEvalContext) -> Vec<UnitCellContext> {
         match &self {
-            Sequencer::Square(g) => g.to_unit_cell_ctxs(),
-            Sequencer::Rect(g) => g.to_unit_cell_ctxs(),
-            Sequencer::Hex(g) => g.to_unit_cell_ctxs(),
+            Sequencer::Square(g) => g.to_unit_cell_ctxs(ctx),
+            Sequencer::Rect(g) => g.to_unit_cell_ctxs(ctx),
+            Sequencer::Hex(g) => g.to_unit_cell_ctxs(ctx),
         }
     }
 }
@@ -29,7 +31,7 @@ pub struct SimpleHexSequence {
     size: f32,
 }
 impl UnitCellCreator for SimpleHexSequence {
-    fn to_unit_cell_ctxs(&self) -> Vec<UnitCellContext> {
+    fn to_unit_cell_ctxs(&self, _ctx: &UnitCellEvalContext) -> Vec<UnitCellContext> {
         make_grid(self.cols, self.rows, vec2(self.size, self.size), true)
     }
 }
@@ -41,7 +43,7 @@ pub struct SimpleSquareSequence {
     size: f32,
 }
 impl UnitCellCreator for SimpleSquareSequence {
-    fn to_unit_cell_ctxs(&self) -> Vec<UnitCellContext> {
+    fn to_unit_cell_ctxs(&self, _ctx: &UnitCellEvalContext) -> Vec<UnitCellContext> {
         make_grid(self.cols, self.rows, vec2(self.size, self.size), false)
     }
 }
@@ -60,7 +62,7 @@ impl SimpleRectSequence {
 }
 
 impl UnitCellCreator for SimpleRectSequence {
-    fn to_unit_cell_ctxs(&self) -> Vec<UnitCellContext> {
+    fn to_unit_cell_ctxs(&self, _ctx: &UnitCellEvalContext) -> Vec<UnitCellContext> {
         make_grid(self.cols, self.rows, self.size, false)
     }
 }
